requirements.txt
# Python Telegram Bot
python-telegram-bot==20.7

# Google API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP клиенты
aiohttp==3.9.1
requests==2.31.0

# Flask для webhook сервера
flask==3.0.0

# Конфигурация
python-dotenv==1.0.0